+++ argopy/fetchers.py
@@ -5,6 +5,8 @@
 High level helper methods to load Argo data from any source
 The facade should be able to work with all available data access point,
 
+Validity of access points parameters (eg: wmo) is made here, not at the data/index source fetcher level
+
 """
 
 import warnings
@@ -13,8 +15,9 @@ import pandas as pd
 
 from argopy.options import OPTIONS, _VALIDATORS
 from .errors import InvalidFetcherAccessPoint, InvalidFetcher
-from .utilities import list_available_data_src, list_available_index_src, is_box
+from .utilities import list_available_data_src, list_available_index_src, is_box, is_indexbox, check_wmo
 from .plotters import plot_trajectory, bar_plot
+
 
 AVAILABLE_DATA_SOURCES = list_available_data_src()
 AVAILABLE_INDEX_SOURCES = list_available_index_src()
@@ -211,14 +214,16 @@ class ArgoDataFetcher:
 
         Parameters
         ----------
-        wmo: list(int)
-            Define the list of Argo floats to load data for. This is a list of integers with WMO numbers.
+        wmo: int, list(int)
+            Define the list of Argo floats to load data for. This is a list of integers with WMO float identifiers.
+            WMO is the World Meteorological Organization.
 
         Returns
         -------
         :class:`argopy.fetchers.ArgoDataFetcher.float`
             A data source fetcher for all float profiles
         """
+        wmo = check_wmo(wmo)  # Check and return a valid list of WMOs
         if "CYC" in kw or "cyc" in kw:
             raise TypeError(
                 "float() got an unexpected keyword argument 'cyc'. Use 'profile' access "
@@ -246,8 +251,9 @@ class ArgoDataFetcher:
 
         Parameters
         ----------
-        wmo: list(int)
-            Define the list of Argo floats to load data for. This is a list of integers with WMO numbers.
+        wmo: int, list(int)
+            Define the list of Argo floats to load data for. This is a list of integers with WMO float identifiers.
+            WMO is the World Meteorological Organization.
         cyc: list(int)
             Define the list of cycle numbers to load for each Argo floats listed in ``wmo``.
 
@@ -256,6 +262,7 @@ class ArgoDataFetcher:
         :class:`argopy.fetchers.ArgoDataFetcher.profile`
             A data source fetcher for specific float profiles
         """
+        wmo = check_wmo(wmo)  # Check and return a valid list of WMOs
         self.fetcher = self.Fetchers["profile"](WMO=wmo, CYC=cyc, **self.fetcher_options)
         self._AccessPoint = "profile"  # Register the requested access point
         self._AccessPoint_data = {'wmo': wmo, 'cyc': cyc}  # Register the requested access point data
@@ -590,26 +597,6 @@ class ArgoIndexFetcher:
         return self._index
 
     @checkAccessPoint
-    def profile(self, wmo, cyc):
-        """ Profile index fetcher
-
-        Parameters
-        ----------
-        wmo: list(int)
-            Define the list of Argo floats to load data for. This is a list of integers with WMO numbers.
-        cyc: list(int)
-            Define the list of cycle numbers to load for each Argo floats listed in ``wmo``.
-
-        Returns
-        -------
-        :class:`argopy.fetchers.ArgoIndexFetcher`
-            A index fetcher initialised for specific float profiles
-        """
-        self.fetcher = self.Fetchers["profile"](WMO=wmo, CYC=cyc, **self.fetcher_options)
-        self._AccessPoint = "profile"  # Register the requested access point
-        return self
-
-    @checkAccessPoint
     def float(self, wmo):
         """ Float index fetcher
 
@@ -620,11 +607,34 @@ class ArgoIndexFetcher:
 
         Returns
         -------
-        :class:`argopy.fetchers.ArgoIndexFetcher`
-            A index fetcher initialised for all float profiles
+        :class:`argopy.fetchers.ArgoIndexFetcher.float`
+            An index source fetcher for all float profiles index
         """
+        wmo = check_wmo(wmo)  # Check and return a valid list of WMOs
         self.fetcher = self.Fetchers["float"](WMO=wmo, **self.fetcher_options)
         self._AccessPoint = "float"  # Register the requested access point
+        return self
+
+    @checkAccessPoint
+    def profile(self, wmo, cyc):
+        """ Profile index fetcher
+
+            Parameters
+            ----------
+            wmo: int, list(int)
+                Define the list of Argo floats to load index for. This is a list of integers with WMO float identifiers.
+                WMO is the World Meteorological Organization.
+            cyc: list(int)
+                Define the list of cycle numbers to load for each Argo floats listed in ``wmo``.
+
+            Returns
+            -------
+            :class:`argopy.fetchers.ArgoIndexFetcher`
+                A index fetcher initialised for specific float profiles
+        """
+        wmo = check_wmo(wmo)  # Check and return a valid list of WMOs
+        self.fetcher = self.Fetchers["profile"](WMO=wmo, CYC=cyc, **self.fetcher_options)
+        self._AccessPoint = "profile"  # Register the requested access point
         return self
 
     @checkAccessPoint
@@ -651,8 +661,8 @@ class ArgoIndexFetcher:
         Warning
         -------
         Note that the box option for an index fetcher does not have pressure bounds, contrary to the data fetcher.
-
         """
+        is_indexbox(box, errors="raise")  # Validate the box definition
         self.fetcher = self.Fetchers["region"](box=box, **self.fetcher_options)
         self._AccessPoint = "region"  # Register the requested access point
         return self